bin: reverse serialization order
Order used to be bottom-up and is now top-down.

Reason for this is that when deserializing, one needs to wait
for the whole buffer to arrive before any deserialization can
take place. Now, the top-level objects arrive first, which allows
for faster deserialization.

The downside is that serialization is slower, since it cannot send
the buffers until all of the objects are serialized, because during
serialization of an object we have to return back to it's head to
write the body's offset.